% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/reduce_log.R
\name{downsample_multiple_by_process}
\alias{downsample_multiple_by_process}
\title{Downsample several columns by process}
\usage{
downsample_multiple_by_process(data, nbins)
}
\arguments{
\item{data}{The data frame to downsample. The first column must be
the "Time" column and there must be a column named "Process".}

\item{nbins}{The (maximum) number of rows of data to return for
  each process.}
}
\value{
A data frame with the same columns as \code{data} but at
  most \code{nbins} rows for each unique process in
  \code{data$Process}.
}
\description{
Downsample a data frame with > 2 columns with the LTTB algorithm,
using the Split-Apply-Combine pattern to downsample the data for
each Process in \code{data} separately.
}

