% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/plot_log.R
\name{plot_log}
\alias{plot_log}
\title{Plot resource use}
\usage{
plot_log(data, title = "", file = "", w = 12, h = 6, ...)
}
\arguments{
\item{data}{Data frame created by reading in
\code{monitor.py} log file or downsampling such a data frame}

\item{title}{Title for the plot (optional)}

\item{file}{File to write the plot to (plot to screen if '')}

\item{w,h}{The width and height of the PDF image in inches}

\item{...}{Other options to pass to \code{pdf()}}
}
\description{
Plot the resource use recorded in a log file with \code{monitor.py}
}

