% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/summarize_log.R
\name{summarize_log}
\alias{summarise_log}
\alias{summarize_log}
\title{Summarize resource use}
\usage{
summarize_log(res)
}
\arguments{
\item{res}{Data frame created by reading in \code{monitor.py} log file}
}
\value{
Data frame summarizing the maximum and median values of the
  CPU, RAM, and Disk I/O usage
}
\description{
Summarize the resource use recorded in a log file with \code{monitor.py}
}

